refactor: move 8 views from shared-web to domain boundary packages (Phase 4a)
Move AgentsView, BoardView, BoardStats, ChannelView, DaemonsView,
IssueTimelineView, ModelsView, and RunSessionUiMeta from shared-web
to their respective domain modules' boundary packages. Update imports
in HtmlViews facade, controllers, and test files. Add sharedWebCore
and domainBceDeps to agent, board, config, daemon, gateway, and
issues domain modules.

shared-web reduced from 31 to 23 view files. Remaining views have
multi-domain dependencies that prevent clean single-domain ownership.
