+++ modules/agent-domain/src/main/scala/agent/boundary/AgentsView.scala
@@ -1,4 +1,6 @@
-package shared.web
+package agent.boundary
+
+import shared.web.*
 
 import zio.json.*
