+++ modules/workspace-domain/src/main/scala/workspace/boundary/RunSessionUiMeta.scala
@@ -1,4 +1,4 @@
-package shared.web
+package workspace.boundary
 
 import workspace.entity.RunStatus
